pyproject.toml
[project]
name = "dbread"
version = "0.4.0"
description = "Read-only database MCP proxy for AI — safe SELECT + MongoDB read access with 5-layer defense"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "tvtdev94" },
]
keywords = [
    "mcp",
    "model-context-protocol",
    "claude",
    "claude-code",
    "ai",
    "database",
    "sql",
    "read-only",
    "postgres",
    "mysql",
    "sqlite",
    "security",
    "sqlglot",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]
dependencies = [
    "mcp>=1.0,<2.0",
    "sqlglot>=23.0",
    "sqlalchemy>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tzdata>=2024.1; sys_platform == 'win32'",
]

[project.urls]
Homepage = "https://github.com/tvtdev94/dbread"
Repository = "https://github.com/tvtdev94/dbread"
Issues = "https://github.com/tvtdev94/dbread/issues"
Documentation = "https://github.com/tvtdev94/dbread#readme"

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9"]
mysql = ["pymysql>=1.1"]
mssql = ["pyodbc>=5.0"]
oracle = ["oracledb>=2.0"]
duckdb = ["duckdb>=1.0", "duckdb-engine>=0.13"]
clickhouse = ["clickhouse-sqlalchemy>=0.3"]
mongo = ["pymongo>=4.6,<5"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.3",
    "ruff>=0.4",
]

[project.scripts]
dbread = "dbread.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dbread"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "integration: requires docker / external services",
]

[tool.coverage.run]
parallel = true
source = ["src/dbread"]
branch = false

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
